backend/requirements.txt
django>=4.2
djangorestframework>=3.15
django-cors-headers>=4.3
python-dotenv>=1.0
gunicorn>=21.2
